[GSoC2024] Enhancement of empty list components (#7659)
### Motivation and context
Fixes #7657 Made all of the empty-list components ui consistent.

### Checklist
- [x] I have created a changelog fragment <!-- see top comment in
CHANGELOG.md -->
- [x] I submit my changes into the `develop` branch
- [x] I have linked related issues (see [GitHub docs](

https://help.github.com/en/github/managing-your-work-on-github/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword))

### License

- [x] I submit _my code changes_ under the same [MIT License](
https://github.com/opencv/cvat/blob/develop/LICENSE) that covers the
project.
  Feel free to contact the maintainers if that's a concern.

---------
