+++ changelog.d/20240403_153414_syedbarimnajan_empty_list_components.md
@@ -1,0 +1,4 @@
+### Fixed
+
+- Standardize the alignment of empty-list components
+  (<https://github.com/opencv/cvat/pull/7659>)
