+++ cvat-ui/package.json
@@ -1,6 +1,6 @@
 {
   "name": "cvat-ui",
-  "version": "1.63.5",
+  "version": "1.63.6",
   "description": "CVAT single-page application",
   "main": "src/index.tsx",
   "scripts": {
