+++ cvat-ui/src/components/cloud-storages-page/cloud-storages-list.tsx
@@ -1,4 +1,5 @@
 // Copyright (C) 2021-2022 Intel Corporation
+// Copyright (C) 2022-2024 CVAT.ai Corporation
 //
 // SPDX-License-Identifier: MIT
 
@@ -7,6 +8,7 @@ import Pagination from 'antd/lib/pagination';
 import { Row, Col } from 'antd/lib/grid';
 
 import { CloudStorage } from 'reducers';
+import dimensions from 'utils/dimensions';
 import CloudStorageItemComponent from './cloud-storage-item';
 
 interface Props {
@@ -36,7 +38,7 @@ export default function StoragesList(props: Props): JSX.Element {
     return (
         <>
             <Row justify='center' align='middle'>
-                <Col span={24} className='cvat-cloud-storages-list'>
+                <Col {...dimensions} className='cvat-cloud-storages-list'>
                     {groupedStorages.map(
                         (instances: CloudStorage[]): JSX.Element => (
                             <Row key={instances[0].id} gutter={[8, 8]}>
