+++ cvat-ui/src/components/cloud-storages-page/cloud-storages-page.tsx
@@ -1,4 +1,5 @@
 // Copyright (C) 2021-2022 Intel Corporation
+// Copyright (C) 2022-2024 CVAT.ai Corporation
 //
 // SPDX-License-Identifier: MIT
 
@@ -6,14 +7,15 @@ import './styles.scss';
 import React, { useCallback, useEffect, useState } from 'react';
 import { useHistory } from 'react-router';
 import { useDispatch, useSelector } from 'react-redux';
-import { Row, Col } from 'antd/lib/grid';
+import { Row } from 'antd/lib/grid';
 import Spin from 'antd/lib/spin';
 
-import { CombinedState, Indexable } from 'reducers';
+import { CloudStoragesQuery, CombinedState, Indexable } from 'reducers';
 import { getCloudStoragesAsync } from 'actions/cloud-storage-actions';
 import { updateHistoryFromQuery } from 'components/resource-sorting-filtering';
+import { anySearch } from 'utils/any-search';
 import CloudStoragesListComponent from './cloud-storages-list';
-import EmptyCloudStorageListComponent from './empty-cloud-storages-list';
+import EmptyListComponent from './empty-list';
 import TopBarComponent from './top-bar';
 
 export default function StoragesPageComponent(): JSX.Element {
@@ -57,15 +59,8 @@ export default function StoragesPageComponent(): JSX.Element {
         [query],
     );
 
-    const dimensions = {
-        md: 22,
-        lg: 18,
-        xl: 16,
-        xxl: 16,
-    };
+    const isAnySearch = anySearch<CloudStoragesQuery>(query);
 
-    const anySearch = Object.keys(query)
-        .some((value: string) => value !== 'page' && (query as any)[value] !== null);
     const content = current.length ? (
         <CloudStoragesListComponent
             totalCount={totalCount}
@@ -74,48 +69,46 @@ export default function StoragesPageComponent(): JSX.Element {
             onChangePage={onChangePage}
         />
     ) : (
-        <EmptyCloudStorageListComponent notFound={anySearch} />
+        <EmptyListComponent notFound={isAnySearch} />
     );
 
     return (
-        <Row className='cvat-cloud-storages-page' justify='center' align='top'>
-            <Col {...dimensions}>
-                <TopBarComponent
-                    onApplySearch={(_search: string | null) => {
-                        dispatch(
-                            getCloudStoragesAsync({
-                                ...query,
-                                search: _search,
-                                page: 1,
-                            }),
-                        );
-                    }}
-                    onApplyFilter={(filter: string | null) => {
-                        dispatch(
-                            getCloudStoragesAsync({
-                                ...query,
-                                filter,
-                                page: 1,
-                            }),
-                        );
-                    }}
-                    onApplySorting={(sorting: string | null) => {
-                        dispatch(
-                            getCloudStoragesAsync({
-                                ...query,
-                                sort: sorting,
-                                page: 1,
-                            }),
-                        );
-                    }}
-                    query={updatedQuery}
-                />
-                { fetching ? (
-                    <Row className='cvat-cloud-storages-page' justify='center' align='middle'>
-                        <Spin size='large' />
-                    </Row>
-                ) : content }
-            </Col>
-        </Row>
+        <div className='cvat-cloud-storages-page'>
+            <TopBarComponent
+                onApplySearch={(_search: string | null) => {
+                    dispatch(
+                        getCloudStoragesAsync({
+                            ...query,
+                            search: _search,
+                            page: 1,
+                        }),
+                    );
+                }}
+                onApplyFilter={(filter: string | null) => {
+                    dispatch(
+                        getCloudStoragesAsync({
+                            ...query,
+                            filter,
+                            page: 1,
+                        }),
+                    );
+                }}
+                onApplySorting={(sorting: string | null) => {
+                    dispatch(
+                        getCloudStoragesAsync({
+                            ...query,
+                            sort: sorting,
+                            page: 1,
+                        }),
+                    );
+                }}
+                query={updatedQuery}
+            />
+            { fetching ? (
+                <Row className='cvat-cloud-storages-page' justify='center' align='middle'>
+                    <Spin size='large' />
+                </Row>
+            ) : content }
+        </div>
     );
 }
