+++ cvat-ui/src/components/cloud-storages-page/styles.scss
@@ -1,13 +1,15 @@
 // Copyright (C) 2021-2022 Intel Corporation
-// Copyright (C) 2022 CVAT.ai Corporation
+// Copyright (C) 2022-2024 CVAT.ai Corporation
 //
 // SPDX-License-Identifier: MIT
 
-@import '../../base.scss';
+@import '../../base';
 
 .cvat-cloud-storages-page {
-    padding: $grid-unit-size * 2;
+    padding-top: $grid-unit-size * 2;
+    padding-bottom: $grid-unit-size;
     height: 100%;
+    width: 100%;
     overflow: auto;
 
     .ant-spin {
@@ -56,7 +58,14 @@
 
 .cvat-cloud-storages-list,
 .cvat-empty-cloud-storages-list {
-    margin-top: $grid-unit-size * 2;
+    margin-top: $grid-unit-size;
+
+    .ant-empty {
+        top: 50%;
+        left: 50%;
+        position: absolute;
+        transform: translate(-50%, -50%);
+    }
 }
 
 .cvat-cloud-storage-item {
