+++ cvat-ui/src/components/cloud-storages-page/top-bar.tsx
@@ -1,4 +1,5 @@
 // Copyright (C) 2021-2022 Intel Corporation
+// Copyright (C) 2022-2024 CVAT.ai Corporation
 //
 // SPDX-License-Identifier: MIT
 
@@ -11,6 +12,8 @@ import { PlusOutlined } from '@ant-design/icons';
 import { CloudStoragesQuery } from 'reducers';
 import Input from 'antd/lib/input';
 import { SortingComponent, ResourceFilterHOC, defaultVisibility } from 'components/resource-sorting-filtering';
+
+import dimensions from 'utils/dimensions';
 
 import {
     localStorageRecentKeyword, localStorageRecentCapacity,
@@ -36,8 +39,8 @@ export default function StoragesTopBar(props: Props): JSX.Element {
     const [visibility, setVisibility] = useState(defaultVisibility);
 
     return (
-        <Row justify='space-between' align='middle' className='cvat-cloud-storages-list-top-bar'>
-            <Col span={24}>
+        <Row justify='center' align='middle' className='cvat-cloud-storages-list-top-bar'>
+            <Col {...dimensions}>
                 <div className='cvat-cloudstorages-page-filters-wrapper'>
                     <Input.Search
                         enterButton
