+++ cvat-ui/src/components/jobs-page/jobs-page.tsx
@@ -1,5 +1,5 @@
 // Copyright (C) 2022 Intel Corporation
-// Copyright (C) 2023-2024 CVAT.ai Corporation
+// Copyright (C) 2022-2024 CVAT.ai Corporation
 //
 // SPDX-License-Identifier: MIT
 
@@ -10,16 +10,16 @@ import { useDispatch, useSelector } from 'react-redux';
 import Spin from 'antd/lib/spin';
 import { Col, Row } from 'antd/lib/grid';
 import Pagination from 'antd/lib/pagination';
-import Empty from 'antd/lib/empty';
-import Text from 'antd/lib/typography/Text';
 
 import { Job } from 'cvat-core-wrapper';
 import { updateHistoryFromQuery } from 'components/resource-sorting-filtering';
-import { CombinedState, Indexable } from 'reducers';
+import { CombinedState, Indexable, JobsQuery } from 'reducers';
 import { getJobsAsync, updateJobAsync } from 'actions/jobs-actions';
+import { anySearch } from 'utils/any-search';
 
 import TopBarComponent from './top-bar';
 import JobsContentComponent from './jobs-content';
+import EmptyListComponent from './empty-list';
 
 function JobsPageComponent(): JSX.Element {
     const dispatch = useDispatch();
@@ -54,6 +54,8 @@ function JobsPageComponent(): JSX.Element {
         }
     }, [query]);
 
+    const isAnySearch = anySearch<JobsQuery>(query);
+
     const content = count ? (
         <>
             <JobsContentComponent onJobUpdate={onJobUpdate} />
@@ -77,16 +79,7 @@ function JobsPageComponent(): JSX.Element {
             </Row>
         </>
     ) : (
-        <div className='cvat-empty-jobs-list'>
-            <Empty description={(
-                <Row justify='center' align='middle'>
-                    <Col>
-                        <Text>No results matched your search...</Text>
-                    </Col>
-                </Row>
-            )}
-            />
-        </div>
+        <EmptyListComponent notFound={isAnySearch} />
     );
 
     return (
