+++ cvat-ui/src/components/projects-page/empty-list.tsx
@@ -1,4 +1,5 @@
 // Copyright (C) 2020-2022 Intel Corporation
+// Copyright (C) 2022-2024 CVAT.ai Corporation
 //
 // SPDX-License-Identifier: MIT
 
@@ -14,6 +15,7 @@ interface Props {
 
 export default function EmptyListComponent(props: Props): JSX.Element {
     const { notFound } = props;
+
     return (
         <div className='cvat-empty-projects-list'>
             <Empty description={notFound ? (
@@ -22,7 +24,7 @@ export default function EmptyListComponent(props: Props): JSX.Element {
                 <>
                     <Row justify='center' align='middle'>
                         <Col>
-                            <Text strong>No projects created yet ...</Text>
+                            <Text strong>No projects created yet...</Text>
                         </Col>
                     </Row>
                     <Row justify='center' align='middle'>
